component deepDive
