+++ src/components/Cockpit/Cockpit.css
@@ -1,15 +1,16 @@
-.App {
-  text-align: center;
-}
-
-.App button.Red {
+.Cockpit button.Red {
   background-color: red;
 }
 
 .bold {
   font-weight: bold;
 }
-.App button {
+
+.red {
+  color: red;
+}
+
+.Cockpit button {
   border: 1px solid blue;
   padding: 16px;
   background-color: green;
@@ -17,11 +18,13 @@
   color: white;
   cursor: pointer;
 }
-.App button:hover {
+
+.Cockpit button:hover {
   background-color: lightgreen;
   color: black;
 }
-.App button.Red:hover {
+
+.Cockpit button.Red:hover {
   background-color: salmon;
   color: black;
 }
