+++ src/containers/App.css
@@ -1,0 +1,5 @@
+.App {
+  text-align: center;
+}
+
+
