+++ src/containers/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import classes from './App.css';
-import Person from './Person/Person';
-import ErrorBoundary from './ErrorBoundary/ErrorBoundary'
+import Persons from '../components/Persons/Persons';
+import Cockpit from '../components/Cockpit/Cockpit';
 
 class App extends Component {
   state = {
@@ -46,41 +46,25 @@ class App extends Component {
   render() {
 
     let persons = null;
-    let btnClass = '';
-
     if (this.state.showPersons) {
       persons = (
-        <div>
-          {this.state.persons.map((person, index) => {
-            return <ErrorBoundary key={person.id}>
-              <Person
-              click={() => this.deletePersonHandler(index)}
-              name={person.name}
-              age={person.age}
 
-              changed={(event) => this.nameChangedHandler(event, person.id)}
-            /></ErrorBoundary>
-          })}
-
-        </div>
+          <Persons
+            persons={this.state.persons}
+            clicked={this.deletePersonHandler}
+            changed={this.nameChangedHandler}
+          />
       );
-      btnClass = classes.Red;
     }
-    const assignedclasses = [];
-    if (this.state.persons.length <= 2) {
-      assignedclasses.push(classes.red);
-    }
-    if (this.state.persons.length <= 1) {
-      assignedclasses.push(classes.bold);
-    }
+
     return (
 
       <div className={classes.App}>
-        <h1>Hi, I'm a React App </h1>
-        <p className={assignedclasses.join(' ')}>It is really working!</p>
-        <button
-          className={btnClass}
-          onClick={this.togglePersonHandler}>Switch Name</button>
+        <Cockpit
+          showPersons={this.state.showPersons}
+          persons={this.state.persons}
+          clicked={this.togglePersonHandler}
+        />
         {persons}
 
       </div>
